+++ src/main/java/com/millioncook/myapp/dao/AuthenticationDao.java
@@ -12,7 +12,7 @@ public interface AuthenticationDao {
 
 	public boolean registerNewUser(User newUser) throws Exception;
 
-	public User getUserByUserId(Integer userId) throws Exception;
+	public User getUserByUserId(Long userId) throws Exception;
 
 	public String updateNewAddress(UserAddress address) throws Exception;
